resolute-ai-pvt-ltd/task2/requirements.txt
streamlit
torch
transformers
sentence-transformers
chromadb
PyMuPDF   
python-docx   
